[refactor] Move game script into script.py
